pyqmicli: handle missing DNS values
I think to remember that in some tests not all DNS values where returned,
either in IPv4 or IPv6. To avoid any crash, let the parsing of start
network QMI call not fail if some data are missing. An error would be returned
anyway when trying to get this values when needed.