qemurun: make boot partition configurable
Currently, the boot partition is /dev/sda5 by default, but I got this
value from try and error, so it needs to be easier to configure, until I
find a proper way to detect this correctly.

This change adds a ROOT variable which is /dev/sda5 by default, for
backward compatibility with older configuration files, and can be
configured in .cfg defining there the ROOT variable.