qemu: clean up option configuration for qemurun
Currently, there are two semi-identicall calls to qemu for headless and
non-headless configuration, which is ugly and harder to maintain (why
there are different options?).
To avoid this, the option string is composed iteratively, which also
permits to add comments, and passed as variable ($OPTS) to qemu command.

This is nicer, however the -append flag does not like it and skip the
part after the first pace (namely the console=ttyS0), but so far is
fine.