qemurun: workaround for running headless
For some reason, I cannot use the -append flags option in a script. It
seems that the quotes are not done right, so qemu does not understand
the command. To overcome this (temporally) the commandline is not
executed directly, but copied into the clipboard (xclip required), so
that the user can paste it into the shell and execute it. This way the
command works fine.